+++ google-modules/lwis/lwis_device_test.c
@@ -79,7 +79,7 @@ static int lwis_test_register_io(struct lwis_device *lwis_dev, struct lwis_io_en
 	case LWIS_IO_ENTRY_READ_BATCH:
 	case LWIS_IO_ENTRY_READ_BATCH_V2: {
 		struct lwis_io_entry_rw_batch *rw_batch;
-		size_t sum_offset_size;
+		uint64_t sum_offset_size;
 
 		rw_batch = &entry->rw_batch;
 		if (rw_batch->offset >= SCRATCH_TEST_DEV_MEMORY_SIZE ||
@@ -87,8 +87,8 @@ static int lwis_test_register_io(struct lwis_device *lwis_dev, struct lwis_io_en
 				       &sum_offset_size) ||
 		    sum_offset_size > SCRATCH_TEST_DEV_MEMORY_SIZE) {
 			dev_err(test_dev->base_dev.dev,
-				"Read range[offset(%llu) + size_in_bytes(%zu)] exceeds scratch memory (%d)\n",
-				rw_batch->offset, rw_batch->size_in_bytes,
+				"Read range[offset(%llu) + size_in_bytes(%llu)] exceeds scratch memory (%d)\n",
+				rw_batch->offset, (unsigned long long)rw_batch->size_in_bytes,
 				SCRATCH_TEST_DEV_MEMORY_SIZE);
 			return -EINVAL;
 		}
@@ -112,7 +112,7 @@ static int lwis_test_register_io(struct lwis_device *lwis_dev, struct lwis_io_en
 	case LWIS_IO_ENTRY_WRITE_BATCH:
 	case LWIS_IO_ENTRY_WRITE_BATCH_V2: {
 		struct lwis_io_entry_rw_batch *rw_batch;
-		size_t sum_offset_size;
+		uint64_t sum_offset_size;
 
 		rw_batch = &entry->rw_batch;
 		if (rw_batch->offset >= SCRATCH_TEST_DEV_MEMORY_SIZE ||
@@ -120,8 +120,8 @@ static int lwis_test_register_io(struct lwis_device *lwis_dev, struct lwis_io_en
 				       &sum_offset_size) ||
 		    sum_offset_size > SCRATCH_TEST_DEV_MEMORY_SIZE) {
 			dev_err(test_dev->base_dev.dev,
-				"Write range[offset(%llu) + size_in_bytes(%zu)] exceeds scratch memory (%d)\n",
-				rw_batch->offset, rw_batch->size_in_bytes,
+				"Write range[offset(%llu) + size_in_bytes(%llu)] exceeds scratch memory (%d)\n",
+				rw_batch->offset, (unsigned long long)rw_batch->size_in_bytes,
 				SCRATCH_TEST_DEV_MEMORY_SIZE);
 			return -EINVAL;
 		}
